Split herbivores into food and moved grids
It will probably have little effect. Some space will be saves since the
booleans can be packed together better when decoupled from the floats.
Maybe that will help reduce memory in the cache or whatever. One benefit
is that the Grid::fill method I added is optimized to memset the way I
used it.